+++ app-api/src/test/java/org/yechan/api/issueToken/POST_spec.java
@@ -2,9 +2,9 @@ package org.yechan.api.issueToken;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.yechan.EmailGenerator.generateEmail;
-import static org.yechan.PhoneNumberGenerator.generatePhone;
-import static org.yechan.UsernameGenerator.generateUsername;
+import static org.yechan.testdata.EmailGenerator.generateEmail;
+import static org.yechan.testdata.PhoneNumberGenerator.generatePhone;
+import static org.yechan.testdata.UsernameGenerator.generateUsername;
 
 import java.util.Base64;
 import org.junit.jupiter.api.DisplayName;
@@ -12,12 +12,12 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.yechan.TestFixture;
 import org.yechan.config.IntegrationTest;
 import org.yechan.dto.TokenHolder;
 import org.yechan.dto.request.IssueTokenRequest;
 import org.yechan.dto.request.UserRegisterRequest;
 import org.yechan.dto.response.RegisterSuccessResponse;
+import org.yechan.fixture.TestFixture;
 
 @IntegrationTest
 @DisplayName("POST /api/v1/auth/issueToken")
