+++ app-api/src/test/java/org/yechan/config/IntegrationTest.java
@@ -4,7 +4,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
-import org.yechan.TestFixtureConfig;
+import org.yechan.fixture.TestFixtureConfig;
 
 @Target({java.lang.annotation.ElementType.TYPE})
 @Retention(java.lang.annotation.RetentionPolicy.RUNTIME)
