+++ app-api/src/test/java/org/yechan/testdata/PasswordGenerator.java
@@ -1,4 +1,4 @@
-package org.yechan;
+package org.yechan.testdata;
 
 public class PasswordGenerator {
     public static String generatePassword() {
