+++ app-api/src/test/java/org/yechan/testdata/UsernameGenerator.java
@@ -1,4 +1,4 @@
-package org.yechan;
+package org.yechan.testdata;
 
 import java.util.UUID;
